フレンドを解除したときにルームを削除するように修正

+++ app/controllers/relationships_controller.rb
@@ -2,23 +2,35 @@ class RelationshipsController < ApplicationController
 
   def create
     @user = User.find_by(id: params[:friend_id])
-    relationship = current_user.friend(@user)
-    notice = Notice.create(sender_id: current_user.id, reciever_id: @user.id, content: "#{current_user.name}さんとフレンドになりました。")
     request = Friendrequest.find_by(user_id: @user.id, friend_id: current_user.id)
-    @request_id = request.id
-    request.destroy
-    # Room.create(user_id: relationship.user_id, friend_id: relationship.friend_id)
+    # もしフレンド申請が取り消しされていた場合
+    if request == nil
+      #　エラーメッセージを表示する
+    else
+      @request_id = request.id
+      request.destroy
+      relationship = current_user.friend(@user)
+      notice = Notice.create(sender_id: current_user.id, reciever_id: @user.id, content: "#{current_user.name}さんとフレンドになりました。")
+      Room.create(user_id: relationship.user_id, friend_id: relationship.friend_id, relationship_id: relationship.id)
+    end
   end
 
   def destroy
-    relationship = Relationship.find(params[:id])
-    if current_user == relationship.user 
-      @user = relationship.friend
+    relationship = Relationship.find_by(id: params[:id])
+    if relationship == nil
+      user = User.find_by(id: params[:user_id])
+      redirect_to user_path(user)
+    else
+      if current_user == relationship.user 
+        @user = relationship.friend
+      end
+      if current_user == relationship.friend
+        @user = relationship.user
+      end
+      room = Room.find_by(relationship_id: relationship.id)
+      room.destroy
+      relationship.destroy
     end
-    if current_user == relationship.friend
-      @user = relationship.user
-    end
-    relationship.destroy
   end
 
 end
